+++ test/bones_api_security_test.dart
@@ -1,8 +1,11 @@
+@Timeout(Duration(seconds: 180))
 // ignore_for_file: discarded_futures
+import 'dart:isolate';
 import 'dart:typed_data';
 
 import 'package:bones_api/bones_api.dart';
 import 'package:bones_api/bones_api_server.dart';
+import 'package:shared_map/shared_map.dart';
 import 'package:test/test.dart';
 
 final fooSha256 =
@@ -23,13 +26,13 @@ void main() {
   });
 
   group('APISecurity.secureRandom', () {
-    var apiSecurity = _MyAPISecurity();
+    var apiSecurity = _MyAPISecurity(sharedStore: SharedStore.notShared());
     _testSecureRandom(apiSecurity.secureRandom());
   });
 
   group('APISecurity', () {
     test('createToken', () {
-      var apiSecurity = _MyAPISecurity();
+      var apiSecurity = _MyAPISecurity(sharedStore: SharedStore.notShared());
 
       var token = apiSecurity.createToken('foo');
 
@@ -52,7 +55,7 @@ void main() {
     });
 
     test('authenticate', () {
-      var apiSecurity = _MyAPISecurity();
+      var apiSecurity = _MyAPISecurity(sharedStore: SharedStore.notShared());
 
       {
         var credential = APICredential('foo', passwordHash: 'foo');
@@ -86,20 +89,20 @@ void main() {
     });
 
     test('authenticateByRequest', () async {
-      var apiSecurity = _MyAPISecurity();
+      var apiSecurity = _MyAPISecurity(sharedStore: SharedStore.notShared());
 
       {
         var request1 = APIRequest(APIRequestMethod.GET, 'foo');
         request1.credential = APICredential('foo', passwordHash: 'foo');
 
-        expect(apiSecurity.authenticateByRequest(request1), isNotNull);
+        expect(await apiSecurity.authenticateByRequest(request1), isNotNull);
         expect(request1.authentication, isNotNull);
         expect(request1.authentication!.username, equals('foo'));
 
         var request2 = APIRequest(APIRequestMethod.GET, 'foo');
         request2.credential = APICredential('bar', passwordHash: 'xxx');
 
-        expect(apiSecurity.authenticateByRequest(request2), isNull);
+        expect(await apiSecurity.authenticateByRequest(request2), isNull);
         expect(request2.authentication, isNull);
       }
 
@@ -108,7 +111,7 @@ void main() {
         request1.parameters['username'] = 'foo';
         request1.parameters['password'] = 'foo';
 
-        expect(apiSecurity.authenticateByRequest(request1), isNotNull);
+        expect(await apiSecurity.authenticateByRequest(request1), isNotNull);
         expect(request1.authentication, isNotNull);
         expect(request1.authentication!.username, equals('foo'));
 
@@ -116,7 +119,7 @@ void main() {
         request2.parameters['username'] = 'foo';
         request2.parameters['password'] = 'not_foo';
 
-        expect(apiSecurity.authenticateByRequest(request2), isNull);
+        expect(await apiSecurity.authenticateByRequest(request2), isNull);
         expect(request2.authentication, isNull);
       }
 
@@ -143,14 +146,107 @@ void main() {
       }
     });
 
-    test('resolveRequestCredential', () {
-      var apiSecurity = _MyAPISecurity();
+    test('authenticateByRequest (shared + Isolate)', () async {
+      final sharedStore = SharedStore.fromUUID();
+
+      var apiSecurity = _MyAPISecurity(sharedStore: sharedStore);
+
+      {
+        var request1 = APIRequest(APIRequestMethod.GET, 'foo');
+        request1.credential = APICredential('foo', passwordHash: 'foo');
+
+        expect(await apiSecurity.authenticateByRequest(request1), isNotNull);
+        expect(request1.authentication, isNotNull);
+        expect(request1.authentication!.username, equals('foo'));
+
+        var request2 = APIRequest(APIRequestMethod.GET, 'foo');
+        request2.credential = APICredential('bar', passwordHash: 'xxx');
+
+        expect(await apiSecurity.authenticateByRequest(request2), isNull);
+        expect(request2.authentication, isNull);
+      }
+
+      {
+        var request1 = APIRequest(APIRequestMethod.GET, 'foo');
+        request1.parameters['username'] = 'foo';
+        request1.parameters['password'] = 'foo';
+
+        expect(await apiSecurity.authenticateByRequest(request1), isNotNull);
+        expect(request1.authentication, isNotNull);
+        expect(request1.authentication!.username, equals('foo'));
+
+        var request2 = APIRequest(APIRequestMethod.GET, 'foo');
+        request2.parameters['username'] = 'foo';
+        request2.parameters['password'] = 'not_foo';
+
+        expect(await apiSecurity.authenticateByRequest(request2), isNull);
+        expect(request2.authentication, isNull);
+      }
+
+      {
+        var isolateOk = await Isolate.run<bool>(() async {
+          var request1 = APIRequest(APIRequestMethod.GET, 'foo');
+          request1.parameters['username'] = 'foo';
+          request1.parameters['password'] = 'foo';
+
+          var authentication1 =
+              await apiSecurity.authenticateByRequest(request1);
+          if (authentication1 == null) {
+            throw StateError("Null `authentication1`");
+          }
+
+          var requestAuthentication1 = request1.authentication;
+          if (requestAuthentication1 == null) {
+            throw StateError("Null `request1.authentication`");
+          }
+
+          if (requestAuthentication1.username != 'foo') {
+            throw StateError("`request1.authentication!.username` != 'foo'");
+          }
+
+          var request2 = APIRequest(APIRequestMethod.GET, 'foo');
+          request2.parameters['username'] = 'foo';
+          request2.parameters['password'] = 'not_foo';
+
+          var authentication2 =
+              await apiSecurity.authenticateByRequest(request2);
+          if (authentication2 != null) {
+            throw StateError("Expected null `authentication2`");
+          }
+
+          if (request2.authentication != null) {
+            throw StateError("Expected null `request2.authentication`");
+          }
+
+          var request3 = APIRequest(APIRequestMethod.GET, 'foo',
+              credential:
+                  APICredential('foo', token: requestAuthentication1.tokenKey));
+
+          var authentication3 =
+              await apiSecurity.authenticateByRequest(request3);
+          if (authentication3 == null) {
+            throw StateError("Null `authentication3`");
+          }
+
+          if (request3.authentication == null) {
+            throw StateError("Null `request3.authentication`");
+          }
+
+          return true;
+        });
+
+        expect(isolateOk, isTrue);
+      }
+    });
+
+    test('resolveRequestCredential', () async {
+      var apiSecurity = _MyAPISecurity(sharedStore: SharedStore.notShared());
 
       {
         var request = APIRequest(APIRequestMethod.GET, 'foo');
         request.credential = APICredential('foo', passwordHash: 'foo');
 
-        var credentials = apiSecurity.resolveRequestCredentials(request);
+        var credentials = await apiSecurity.resolveRequestCredentials(request);
         expect(credentials, isNotEmpty);
         expect(credentials.length, equals(1));
         expect(credentials[0].username, equals('foo'));
@@ -171,7 +267,7 @@ void main() {
             APIRequest(APIRequestMethod.GET, 'foo', sessionID: 'SID123abc');
         request.credential = APICredential('', token: fooSha256);
 
-        var credentials = apiSecurity.resolveRequestCredentials(request);
+        var credentials = await apiSecurity.resolveRequestCredentials(request);
         expect(credentials, isNotEmpty);
         expect(credentials.length, equals(1));
         expect(credentials[0].username, equals('foo'));
@@ -181,7 +277,7 @@ void main() {
         var request = APIRequest(APIRequestMethod.GET, 'foo',
             parameters: {'user': 'foo', 'pass': '123456'});
 
-        var credentials = apiSecurity.resolveRequestCredentials(request);
+        var credentials = await apiSecurity.resolveRequestCredentials(request);
         expect(credentials, isNotEmpty);
         expect(credentials.length, equals(1));
         expect(credentials[0].username, equals('foo'));
@@ -190,7 +286,7 @@ void main() {
     });
 
     test('doRequestAuthentication', () async {
-      var apiSecurity = _MyAPISecurity();
+      var apiSecurity = _MyAPISecurity(sharedStore: SharedStore.notShared());
 
       {
         var request =
@@ -225,7 +321,7 @@ void main() {
     });
 
     test('resumeAuthenticationByRequest', () {
-      var apiSecurity = _MyAPISecurity();
+      var apiSecurity = _MyAPISecurity(sharedStore: SharedStore.notShared());
 
       {
         var request =
@@ -297,7 +393,7 @@ void main() {
     });
 
     test('APIRouteRule', () {
-      var apiSecurity = _MyAPISecurity();
+      var apiSecurity = _MyAPISecurity(sharedStore: SharedStore.notShared());
 
       {
         expect(APIRoutePublicRule().toJson(), equals({'rule': 'public'}));
@@ -407,6 +503,8 @@ void _testSecureRandom(SecureRandom random) {
 }
 
 class _MyAPISecurity extends APISecurity {
+  _MyAPISecurity({super.sharedStore});
+
   @override
   String generateToken(String username) {
     if (username == 'foo') {