+++ CHANGELOG.md
@@ -25,8 +25,10 @@
 
 - `SQLGenerator.generateCreateTableSQL`: skip annotated hidden fields.
 
-- New `APITokenStore`.
+- New `APITokenStore`:
+  - Shared tokens among `Isolate`s. 
 
+- shared_map: ^1.0.6
 - args_simple: ^1.0.0
 - coverage: ^1.7.1
 - vm_service: ^13.0.0